aura_backend/pyproject.toml
[project]
name = "aura-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncio-mqtt>=0.16.2",
    "beautifulsoup4>=4.13.4",
    "chromadb>=0.5.23",
    "ebooklib>=0.19",
    "faiss-cpu>=1.11.0",
    "faiss-gpu-cu12>=1.11.0",
    "fastapi>=0.115.6",
    "fastmcp>=2.5.2",
    "google-genai>=1.17.0",
    "google-generativeai>=0.8.3",
    "httpx>=0.28.1",
    "mcp>=1.1.0",
    "memvid>=0.1.3",
    "numpy>=1.21.0,<2.0.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pydantic>=2.10.4",
    "pypdf2>=3.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.2",
    "sentence-transformers>=3.3.1",
    "torch==2.7.0",
    "uvicorn>=0.34.0",
    "websockets>=14.1",
]
